feat: * add film dao

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -40,7 +40,7 @@ public class FilmController {
     }
 
     @GetMapping("/popular")
-    public Collection<Film> popular(@RequestParam(defaultValue = "10") long count) {
+    public Collection<Film> popular(@RequestParam(defaultValue = "10") int count) {
         return service.getTenMostPopularFilms(count);
     }
 }
